requirements.txt
Django==3.2.10
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
python-scrapyd-api==2.1.2
requests==2.26.0
python-decouple==3.3
psycopg2-binary==2.8.6
django-on-heroku==1.1.2
gunicorn==20.1.0
whitenoise==5.3.0
